pyproject.toml
[project]
name = "libris-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Inácio Oliveira",email = "inacioigne@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.118.2,<0.119.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncmy (>=0.2.10,<0.3.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "libris_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = 'fastapi dev src/app.py'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_test = 'task lint'
test = 'pytest --cov=src --cov-report=term-missing'
post_test = 'coverage html'
[dependency-groups]
dev = [
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "ruff (>=0.14.0,<0.15.0)"
]
